Move example webpack config under examples/
If we're not using webpack to build libram, this file is going to
confuse people simply by existing.
If we're using webpack to build libram, then this file needs to be
moved somewhere else, and we should put our build config here.
Either way, the example build config doesn't quite belong here.