feat: improve lookup results for players that don't exist
KoLmafia getPlayerId/getPlayerName simply return the input string if a
player does not exist. This leads to strange behaviour in
getPlayerFromIdOrName where it returns `{name:0, id: null}` or
`{name:"<id>", id: <id>}`. This is very inconvenient to handle,
especially if the code expects that a user might not exist (because it's
handling user input for example).

Solve by adding getPlayerIdFromName and getPlayerNameFromId that can be
configured to either throw an error or return null if the player does
not exist.
Add the same parameter to getPlayerFromIdOrName, defaulting to throw an
exception, in order not to break signature compatibility with existing
code.